pyproject.toml
[project]
name = "trip-recommender"
version = "0.1.0"
description = ""
authors = [
    {name = "Hasaan"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "openai (>=1.63.2,<2.0.0)",
    "promptlayer (>=1.0.38,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"